added logic to avoid reptition of questions, score tracking, rearrangement of game components, and game over view
