+++ src/Components/GameOverDisplay.js
@@ -1,0 +1,8 @@
+
+const GameOverDisplay = () => {
+    return (
+        <p>Game Over!</p>
+    )
+}
+
+export default GameOverDisplay
